Kernel version 4.0
The Kernel based on priority based algorithm with the ability to  handle multiple tasks with the same priority. The Kernel uses the First delay concept  for best CPU utilization and prevent overhead CPU load.
The Kernel has some objects like binary and count semaphores with the ability to choose either semaphore initially exist or not, this help in wait to signal concept.
the Kernel includes queues to handle the synchronization. The queue is a generic which help you to select your desired length , data types with check on full or empty features.
The kernel has some services like suspend Task to send it to the block state , Resume Task to back to the ready state and delete task to send it to the dormant state.